Fix memory leak due to BIO_NOCLOSE on memory BIO
When BIO_NOCLOSE is set on a memory BIO then the underlying BUF_MEM
structure will not be freed when the BIO is freed.

Originally BIO_NOCLOSE was set when the output was done to stdout to
prevent closing the stream. Seemingly the flag was then carried through
all following refactorings.